Scoring
This commit adds a simple scoring system. You have 60 seconds to
enter your guess (the period may be changed using the TIMEOUT
define in common.h). If your guess is wrong 60 (or whatever value
TIMEOUT is set to) points are added to your score. If the guess
was right only the time used to enter your guess gets added to
your score.

The player with the lowest score is the winner.

At the moment the player gets no feedback how much time has
passed. But this won't change until the UI has been
reimplemented; this is just a prototype ;)

 * add simple scoring
 * show scoreboard

Closes GH-1.