Implement pass-through options for ogg123
The old parameter `-o` has been renamed to `-O`. The new
behaviour of `-o` is to take a string which will be passed on to
all `ogg123` invocations. This means you will be able to use

        oggquiz -o "-k 1:00"

to skip the first minute in all tracks.

CAVEAT EMPTOR: The argument of `-o` is just a string to `oggquiz`
without any semantic meaning. You might run into trouble if you
skip the first minute of a song which has a playing time less
than a minute.

Closes GH-6.