Refactor ui_model handling in oggquiz.c
`struct ui_model` was held in a static variable and was never freed. Resolve
this issue by implementing dedicated functions to create and destroy the
ui_model. Manage the lifetime of ui_model in main function.

While here move the initial seeding of the random number generator to the main
function eliminating the need for any static variables.

 * create dedicated functions to create and destroy `struct ui_model`
 * manage lifetime of `struct ui_model` in main function
 * seed random number generator in main function